requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
pyTelegramBotAPI==4.3.1
requests==2.27.1
urllib3==1.26.8
python-dotenv==0.19.2
loguru==0.5.3
